requirements.txt
streamlit
streamlit_option_menu
google-generativeai
streamlit_folium
pandas
altair
inference-sdk
rasterio
shapely
pyproj
torch
authlib
fastapi
uvicorn
requests
plotly
folium
opencv-python
pillow
numpy
scikit-learn
matplotlib
seaborn